Where do I start..
1. Updated Login page
2. Added SignUp page that uses the create user post api call to create a new user. ***NEEDS FIXING*** Currently, the user_id's are getting mixed up. I created a new user called "test123" and now all of the journals associated with "seth" belong to test123. When I login as Seth, there are no journals.
3. "Entry" page didn't update the markdown content when you click on edit. I tried updating this and found a way to get it to immediately show the changes. There is still an issue where it won't update the context, so if you go to a different journal entry it will show the text of the previous entry you were at.

+++ src/views/journalsAll/index.tsx
@@ -113,8 +113,11 @@ function JournalsAll () {
         }
     };
 
-    const handleJournalDelete = async (id: string) => {
-        await axios.delete(`${context.API_BASE_URL}/api/v1/journal/${id}`, config)
+    const handleJournalDelete = async (id: string) => { 
+        if (window.confirm("Are you sure you want to delete this journal? This cannot be undone.")) { //confirm the deletion of the journal
+            //if yes, delete the journal
+            await axios.delete(`${context.API_BASE_URL}/api/v1/journal/${id}`, config)
+        }
         getJournals();
     };
